fix add opacity method
- replace int 0 to 255 with float 0.0 to 1.0 like Electron and macOS
  API.
- remove WS_EX_LAYERED if opacity is 1.0.
- replace SetWindowLong with SetWindowLongPtr for better compatibility

+++ linux/java/org/jetbrains/jwm/WindowX11.java
@@ -89,7 +89,7 @@ public class WindowX11 extends Window {
     }
 
     @Override
-    public Window setOpacity(int opacity) {
+    public Window setOpacity(float opacity) {
         // TODO: impl me!
         return this;
     }
