+++ shared/java/org/jetbrains/jwm/Window.java
@@ -59,10 +59,11 @@ public abstract class Window extends RefCounted {
     public abstract Window setMouseCursor(MouseCursor cursor);
     
     /**
-    *  Sets window opacity (0 - 255)
+    *  Sets window opacity (0.0 - 1.0)
+    *  If the opacity is outside the range, it will be rounded to 0.0 to 1.0.
      */
     @NotNull @Contract("-> this")
-    public abstract Window setOpacity(int opacity);
+    public abstract Window setOpacity(float opacity);
 
     public abstract Screen getScreen();
