+++ windows/java/org/jetbrains/jwm/WindowWin32.java
@@ -82,7 +82,7 @@ public class WindowWin32 extends Window {
     }
 
     @Override
-    public Window setOpacity(int opacity) {
+    public Window setOpacity(float opacity) {
         assert _onUIThread();
         _nSetOpacity(opacity);
         return this;
@@ -140,7 +140,7 @@ public class WindowWin32 extends Window {
     @ApiStatus.Internal public native void _nSetContentSize(int width, int height);
     @ApiStatus.Internal public native void _nSetTitle(String title);
     @ApiStatus.Internal public native void _nSetIcon(String iconPath);
-    @ApiStatus.Internal public native void _nSetOpacity(int opacity);
+    @ApiStatus.Internal public native void _nSetOpacity(float opacity);
     @ApiStatus.Internal public native void _nSetMouseCursor(int cursorId);
     @ApiStatus.Internal public native Screen _nGetScreen();
     @ApiStatus.Internal public native void _nRequestFrame();
